[v0.1] Basic Count Statistics - Item Counter implemented
Function itemCounts returns a Map with the count of each element in the array received as parameter. For every element in the array, if it already exists in the Map, the code store it's current value to be incrementedin the Map. Otherwise, it is stored with a initial value of 1